requirements.txt
python-telegram-bot==20.3
requests==2.28.2
python-dotenv==0.21.0